etl/postgres_to_es/app/requirements.txt
annotated-types==0.6.0
certifi==2023.7.22
elastic-transport==8.4.1
elasticsearch==8.10.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.8.0
urllib3==1.26.17
